Add status pill back
